Cleanup config_room module
 - Extract my room methods to config_room_my
 - Extract and rename in_queue and get_part_config to config_room_utils
 - Introduce config_room_cached
